+++ src/common/constants/constants.js
@@ -1,7 +1,19 @@
 export const ORGANIZATION_STATUS_ACTIVE = 'Active';
 
 export const PRODUCT_ID_TYPE = {
+  asin: 'ASIN',
+  coden: 'CODEN',
+  doi: 'DOI',
+  gpoItemNumber: 'GPO item number',
   isbn: 'ISBN',
+  ismn: 'ISMN',
+  issn: 'ISSN',
+  lccn: 'LCCN',
+  publisherOrDistributorNumber: 'Publisher or distributor number',
+  reportNumber: 'Report number',
+  standardTechnicalReportNumber: 'Standard technical report number',
+  upc: 'UPC',
+  urn: 'URN',
 };
 export const QUALIFIER_SEPARATOR = ' ';
 
@@ -48,3 +60,5 @@ export const FIELD_ARRAY_ITEM_IDENTIFIER_KEY = '__key__';
 export const OKAPI_TENANT_HEADER = 'X-Okapi-Tenant';
 
 export const HIDDEN_ORDER_FIELDS_FOR_VERSION_HISTORY = ['nextPolNumber'];
+
+export const CUSTOM_FIELDS_BACKEND_MODULE_NAME = 'Orders CRUD module';