+++ src/settings/OrderTemplates/OrderTemplateView/OrderTemplateView.js
@@ -23,13 +23,15 @@ import {
   Pane,
   Row,
 } from '@folio/stripes/components';
-
+import { ViewCustomFieldsRecord } from '@folio/stripes/smart-components';
 import {
   FundDistributionView,
   ORDER_FORMATS,
   handleKeyCommand,
 } from '@folio/stripes-acq-components';
 
+import { CUSTOM_FIELDS_BACKEND_MODULE_NAME } from '../../../common/constants';
+import { isOngoing } from '../../../common/POFields';
 import { PODetailsView } from '../../../components/PurchaseOrder/PODetails';
 import { SummaryView } from '../../../components/PurchaseOrder/Summary';
 import { OngoingOrderInfoView } from '../../../components/PurchaseOrder/OngoingOrderInfo';
@@ -46,7 +48,6 @@ import {
   ERESOURCES,
   PHRESOURCES,
 } from '../../../components/POLine/const';
-import { isOngoing } from '../../../common/POFields';
 import {
   ORDER_TEMPLATES_ACCORDION_TITLES,
   ORDER_TEMPLATES_ACCORDION,
@@ -179,6 +180,7 @@ class OrderTemplateView extends Component {
     const showPhresources = PHRESOURCES.includes(orderFormat);
     const showOther = orderFormat === ORDER_FORMATS.other;
     const orderType = get(orderTemplate, 'orderType');
+    const customFieldsValues = get(orderTemplate, 'customFields', {});
 
     const estimatedPrice = get(orderTemplate, ['cost', 'poLineEstimatedPrice'], 0);
     const fundDistributions = get(orderTemplate, 'fundDistribution');
@@ -199,6 +201,7 @@ class OrderTemplateView extends Component {
       [ORDER_TEMPLATES_ACCORDION.POL_OTHER_RESOURCES]: false,
       [ORDER_TEMPLATES_ACCORDION.POL_LOCATION]: false,
       [ORDER_TEMPLATES_ACCORDION.POL_TAGS]: false,
+      [ORDER_TEMPLATES_ACCORDION.PO_CUSTOM_FIELDS]: false,
     };
     const shortcuts = [
       {
@@ -410,6 +413,20 @@ class OrderTemplateView extends Component {
                     >
                       <OrderTemplateTagsView tags={get(orderTemplate, 'polTags.tagList')} />
                     </Accordion>
+
+                    <ViewCustomFieldsRecord
+                      accordionId={ORDER_TEMPLATES_ACCORDION.PO_CUSTOM_FIELDS}
+                      backendModuleName={CUSTOM_FIELDS_BACKEND_MODULE_NAME}
+                      customFieldsValues={customFieldsValues}
+                      entityType="purchase_order"
+                    />
+
+                    <ViewCustomFieldsRecord
+                      accordionId={ORDER_TEMPLATES_ACCORDION.POL_CUSTOM_FIELDS}
+                      backendModuleName={CUSTOM_FIELDS_BACKEND_MODULE_NAME}
+                      customFieldsValues={customFieldsValues}
+                      entityType="po_line"
+                    />
                   </AccordionSet>
                 </Col>
               </Row>