+++ CHANGELOG.md
@@ -13,9 +13,12 @@
 * Do not show special fields in "Version history" for PO and POL. Refs UIOR-1117.
 * Add validation for the `claimingInterval` field. Refs UIOR-1192.
 * *BREAKING* Add settings components for custom fields. Refs UIOR-1177.
+* Add Custom Fields to View, Edit, Create components of Orders, Order Lines and Order Templates. Refs UIOR-1195.
 * Bump up okapi interfaces for `pieces` (2.0 3.0). Refs UIOR-1212.
+* Invoice line Subscription fields are not populated correctly. Refs UIOR-1220.
 * Opening and editing POs with location-restricted funds (FE). Refs UIOR-1184.
 * Location-restricted funds in POLs. Refs UIOR-1166.
+* Add "LCCN" to the list of valid product identifier types. Refs UIOR-1218.
 * Allow order template to hide Currency and exchange rate fields. Refs UIOR-1214.
 
 ## [5.0.1](https://github.com/folio-org/ui-orders/tree/v5.0.1) (2023-11-08)